backend/requirements.txt
Django==4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pymongo>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
